装机单 显示 bug修复！

+++ app/src/main/java/qx/app/freight/qxappfreight/bean/response/DeclareWaybillBean.java
@@ -83,17 +83,17 @@ public class DeclareWaybillBean implements Serializable {
     /**
      * 重量
      */
-    private int billingWeight;
+    private double billingWeight;
 
     /**
      * 总重量
      */
-    private int totalWeight;
+    private double totalWeight;
 
     /**
      * 总体积
      */
-    private int totalVolume;
+    private double totalVolume;
 
     /**
      * 冷藏温度
